+++ resource/doc.go
@@ -1,0 +1,2 @@
+// Package resource provides commonly used resources (in the ECS sense).
+package resource